requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.2
django-environ==0.11.2
psycopg2-binary==2.9.9
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
holidays==0.60